homework_07/requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
psycopg2-binary==2.9.3
Werkzeug==2.0.3
SQLAlchemy==1.4.23
gunicorn==20.1.0
python-dotenv
asyncpg
aiohttp

